feat: migrate database from SQLite to PostgreSQL and enhance migration scripts

+++ prisma/migrations/20260519120000_add_current_cash/migration.sql
@@ -1,2 +1,2 @@
 -- Add currentCash to FinancialProfile for balance tracking.
-ALTER TABLE "FinancialProfile" ADD COLUMN "currentCash" REAL NOT NULL DEFAULT 0;
+ALTER TABLE "FinancialProfile" ADD COLUMN "currentCash" DOUBLE PRECISION NOT NULL DEFAULT 0;
